Split facet components to own files
